+++ src/scheduler/scheduler.tsx
@@ -3,56 +3,16 @@ import { Scheduler } from "@bitnoi.se/react-scheduler";
 import { useCallback, useState } from "react";
 import moment from "moment";
 
-import Backdrop from '@mui/material/Backdrop';
-import Box from '@mui/material/Box';
-import Modal from '@mui/material/Modal';
-import Fade from '@mui/material/Fade';
-import Typography from '@mui/material/Typography';
-import { Button, ButtonGroup, FormControl, TextField } from "@mui/material";
-import { DateTimePicker } from "@mui/x-date-pickers";
+import { TaskModal } from "./taskmodal";
+import { TaskData, MockedSchedule } from "./types";
 
-const style = {
-  position: 'absolute' as const,
-  top: '50%',
-  left: '50%',
-  transform: 'translate(-50%, -50%)',
-  width: 300,
-  bgcolor: 'background.paper',
-  border: '2px solid ',
-  boxShadow: 24,
-  p: 3,
-  justifyContent: "center",
-  alignContent: "space-between",
-  // height: 700,
-};
 
-type Data = {
-  id: string,
-  startDate: Date,
-  endDate: Date,
-  occupancy: number,
-  title: string,
-  subtitle: string,
-  description: string,
-  bgColor: string,
-}
-
-type MockedSchedule = {
-  id: string,
-  label: {
-    icon: string,
-    title: string,
-    subtitle: string,
-  },
-  data: Data[]
-
-}
 export function Component() {
   const [filterButtonState, setFilterButtonState] = useState(0);
 
   //modal
   const [open, setOpen] = useState(false);
-  const [selectedItem, setSelectedItem] = useState<Data>();
+  const [selectedItem, setSelectedItem] = useState<TaskData>();
   const handleOpen = () => setOpen(true);
   const handleClose = () => setOpen(false);
 
@@ -83,7 +43,7 @@ export function Component() {
   const filteredMockedSchedulerData = mockedSchedulerData.map((person: MockedSchedule) => ({
     ...person,
     data: person.data.filter(
-      (project: Data) =>
+      (project: TaskData) =>
         // we use "dayjs" for date calculations, but feel free to use library of your choice
         moment(project.startDate).isBetween(range.startDate, range.endDate) ||
         moment(project.endDate).isBetween(range.startDate, range.endDate) ||
@@ -100,7 +60,7 @@ export function Component() {
         // onTileClick={(clickedResource) => console.log(clickedResource)}
         onTileClick={(item) => {
           console.log(item)
-          setSelectedItem(item as Data);
+          setSelectedItem(item as TaskData);
           handleOpen();
         }}
         onItemClick={(item) => console.log(item)}
@@ -119,82 +79,11 @@ export function Component() {
           defaultTheme: "dark"
         }}
       />
-      <Modal
-        aria-labelledby="transition-modal-title"
-        aria-describedby="transition-modal-description"
+      <TaskModal
         open={open}
-        onClose={handleClose}
-        closeAfterTransition
-        slots={{ backdrop: Backdrop }}
-        slotProps={{
-          backdrop: {
-            timeout: 500,
-          },
-        }}
-      >
-        <Fade in={open}>
-          <Box component="form" sx={style}>
-            <Typography id="transition-modal-title" variant="h6" component="h2" align="center" >
-              Details on the tile
-            </Typography>
-            <FormControl sx={{ width: '100%' }}>
-
-              <TextField
-                sx={{ my: 1 }}
-                required
-                id="outlined-required"
-                label="Nom"
-                defaultValue={selectedItem?.title}
-              />
-              <TextField
-                sx={{ my: 1 }}
-                required
-                id="outlined-required"
-                label="ID"
-                defaultValue={selectedItem?.id}
-              />
-              <TextField
-                sx={{ my: 1 }}
-                required
-                id="outlined-required"
-                label="subtitle"
-                defaultValue={selectedItem?.subtitle}
-              />
-              <TextField
-                sx={{ my: 1 }}
-                required
-                id="outlined-required"
-                label="description"
-                defaultValue={selectedItem?.description}
-              />
-
-              <TextField
-                sx={{ my: 1 }}
-                required
-                id="outlined-required"
-                label="Couleur"
-                defaultValue={selectedItem?.bgColor}
-              />
-              <DateTimePicker
-                sx={{ my: 1 }}
-                label="Date de début"
-                defaultValue={moment(selectedItem?.startDate)}
-              />
-
-              <DateTimePicker
-                sx={{ my: 1 }}
-                label="Date de fin"
-                defaultValue={moment(selectedItem?.endDate)}
-              />
-              <ButtonGroup variant="contained" aria-label="" sx={{ justifyContent: "end", mt: 2 }}>
-                <Button>Reset</Button>
-                <Button>Apply</Button>
-              </ButtonGroup>
-            </FormControl>
-          </Box>
-        </Fade>
-      </Modal>
-
+        handleClose={handleClose}
+        selectedItem={selectedItem}
+      />
     </div>
   );
 }
